Specification providers corrections

+++ backend/src/main/java/com/norwayyachtbrockers/repository/specification/yacht/provider/DraftDepthSpecificationProvider.java
@@ -3,6 +3,7 @@ package com.norwayyachtbrockers.repository.specification.yacht.provider;
 import com.norwayyachtbrockers.model.Yacht;
 import com.norwayyachtbrockers.repository.specification.SpecificationProvider;
 import com.norwayyachtbrockers.util.YachtSpecificationUtil;
+import java.math.BigDecimal;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Component;
 
@@ -16,7 +17,7 @@ public class DraftDepthSpecificationProvider implements SpecificationProvider<Ya
     @Override
     public Specification<Yacht> getSpecification(Object param) {
         return YachtSpecificationUtil.getSpecificationInRangeOrElseThrow(
-                (Integer[]) param, "yachtModel", getKey()
+                (BigDecimal[]) param, "yachtModel", getKey()
         );
     }
 }
